+++ tests/Integration/ServiceAvailabilityTest.php
@@ -71,6 +71,7 @@ class ServiceAvailabilityTest extends IntegrationTestCase
             [\OxidEsales\MediaLibrary\Media\Settings\FallbackMediaSettingsInterface::class],
             [\OxidEsales\MediaLibrary\Media\Service\FallbackMediaResourceInterface::class],
             [\OxidEsales\MediaLibrary\Media\Service\FallbackMediaSeederInterface::class],
+            [\OxidEsales\MediaLibrary\Media\Service\MediaDeletionPolicyServiceInterface::class],
 
             [\OxidEsales\MediaLibrary\Media\Twig\MediaDataExtension::class],
             [\OxidEsales\MediaLibrary\Media\Twig\MediaDataLogicInterface::class],